Remove add event, extract resources, add entry for volunteering
